Added unit tests for the use case itself.
It won't be able to delete the Knowledges Section

+++ api/ExpressedRealms.Expressions.UseCases/ExpressionTextSections/DeleteTextSection/DeleteTextSectionUseCase.cs
@@ -21,7 +21,14 @@ internal sealed class DeleteTextSectionUseCase(
 
         if (result.IsFailed)
             return Result.Fail(result.Errors);
+        
+        var expressionSection = await repository.GetExpressionSectionForDeletion(model.ExpressionId, model.Id);
 
+        if (expressionSection!.SectionType.Name == "Knowledges Section")
+        {
+            return Result.Fail("You cannot delete the systems knowledge section.");
+        }
+        
         await repository.DeleteExpressionTextSectionAsync(model.ExpressionId, model.Id);
 
         return Result.Ok();
